Data Pathway notes/weather-etl/requirements.txt
requests
pandas
sqlalchemy
mysql-connector-python
python-dotenv